Added functionality for Edit map

+++ app/src/main/java/com/app/risk/utility/MapReader.java
@@ -124,7 +124,6 @@ public class MapReader {
                             if (tempContinent != null) {
 
                                 if (countryGameMapList != null && countryGameMapList.containsKey(words[0])) {
-
                                     countryGameMapList.get(words[0]).getFromCountry().setBelongsToContinent(tempContinent);
                                 } else {
                                     countryGameMapList.put(words[0], new GameMap(new Country(words[0], tempContinent)));
@@ -133,8 +132,8 @@ public class MapReader {
                                 GameMap gameMapForFinalList = countryGameMapList.get(words[0]);
 
                                 gameMapForFinalList.setFromCountry(countryGameMapList.get(words[0]).getFromCountry());
-                                gameMapForFinalList.setCoordinateX(Integer.parseInt(words[1]));
-                                gameMapForFinalList.setCoordinateY(Integer.parseInt(words[2]));
+                                gameMapForFinalList.setCoordinateX(Float.parseFloat(words[1]));
+                                gameMapForFinalList.setCoordinateY(Float.parseFloat(words[2]));
                                 gameMapForFinalList.setConnectedToCountries(setAdjacentCountriesList(words));
 
                                 finalGameMapList.add(gameMapForFinalList);
